Get login working against the actual server.  The tests are incorrect, hence broken, and need to be extended, anyway.
A few constants are included in the AccountModel defaults object.  Some of
these need to be extracted into a separate data source, like a file, that
can be specific to a particular brand packaging.

The account model instance now acquires a bunch of settings on successful
login, and it is completely .clear()ed on logout.