Revise the login process so it respects the username that the server infers from the login name that the user entered.
This is necessary because the server resolves disregards the case of
alphabetic characters in the given username to find the matching account
username, and then returns a login url with a b32 encoding of the actual
account username.

So we must use the returned b32 encoding to determine the actual account
username, and work with that.

It turns out that our tests uncovered a problem with our handling of the
"location:" phase of the login protocol, which is repaired here. It also
turns out that that the server never actually behaves the way those tests
do, but the API documentation (the
[https://spideroak.com/faq/questions/37/how_do_i_use_the_spideroak_web_api/ FAQ entry])
doesn't preclude it, so the routine is now dealing with it consistently.