Set storageRootURL correctly, for getStorageURL().
Also, include the concluding login response data, so we can record the
webRootURL for the account.  We will eventually need to use that,
when (if?) we offer a "visit account via the web" link. (Currently, the API
can return both complete URLs, with the protocol and domain, and ones
without.  That may change soon, to only complete URLs.)