Implement model-less view that consolidates a listing of share rooms currently being visited by the app and the list of share rooms associated with the account (when there is a logged in account).
This is not yet working, but further along.  Further along in the right
direction is the question; this checkin will enable some review.

The current, outstanding problem is putting the visited set of share rooms
in a collection still requires that the collection have a URL - but it's
just a local collection of share rooms, the ones being visited.

Rename ShareRoomsView.js to ShareRoomsViews.js, since it's, significantly,
more than one.  Adjust grunt concat accordingly.