Finally have resolved the proper base mySharesRootURL:
    https://dom.ain/share/

The proper individual share rooms url is:

    share_id_b32 + "/" + shareroom.room_key + "/"

In order to use mySharesRootURL for the respective share rooms, while using
mySharesListURL for fetching the encompassing list of the share rooms, we
change the collection's .url in myShareRoomsListView, after the list has
been fetched and parsed.

With my rudimentary grasp of Backbone way of doing things, I have to figure
out how to implement what is probably the proper answer, which would be to
use a different collection for the list of the account's share rooms than
for the individual share rooms, themselves.  I will look at implementing
that next.

In the meanwhile, this works and should be sound.