Implement a reasonable ShareRoomModel: - .initialize() so each instance has a suitable .url and .get("url") - Parse the Web API data properly
The VisitedShareRoomRootItemModels don't actually need urls, so remove the
initialize() method.  Instead, make getShareRoom (almost) properly
instantiate a model.

The remaining problem is that the url used to actually fetch the share room
is wrong - it is derived from the current page location for the base, plus
the share room url.  I think once I track down why that's happening this
will work.