Implement working Public ShareRooms 'add' and elective remembering of rooms across sessions.
Move the fetch for added models from the collection fetch to the new add
handler.  Implement a remove, that also unremembers any removals that were
remembered.

Handle attempts to add non-existent rooms, in particular not including the
bad stub in the collection.  For now, we post an alert indicating the
mistake.  That should probably become a toast.

Set a default for the account model's b32username attribute, so we can
depend on it being cleared when logged out.  (This might not matter, since
the app may be structure to not allow any real operation without being
logged in, but public share rooms will work, with a remembered set for the
anonymous account different than that of other accounts.)

Don't show a separate "Loading" sign for the Public Share Rooms
section.  We're starting the publics before the My's, so the Loading sign
for the My section should suffice. And I don't have to resolve clearing the
Loading sign from the public section when there are no remembered
ones. (But I may need to do so for no My ones!  Will check that next.)