Working Add Public Share Room UI
However, the actual pop up form for adding one has to go. In my head it needed to be like the login view so it could be used instead of the login when not logged in, but of course that's not really how it works. When not logged in we will be in the share rooms lists view an so therefore the add form just needs to be another pushed subview, not a model over the top of everything view.