Whoops - some changes missed in last checkin.  Same checkin message:
Add a settings view template for the server setting, and start to implement
the alternate server setting.

I still have to get the form right, and then implement a proper handler -
maybe viewChanging?  And then plug the spiderOakApp.config.settings value
in to the AccountModel, and ensure that a logout happens if the setting is
changed during a login.

(Also, rename the settingsAccountViewTemplate.html from the typo, setingsAccountViewTemplate.html.)