Implement a view by which the primary server configuration is changed.
If the change happens during an authenticated session, the session is
logged out and the UI is reset.

If the change happens outside of any authentication, the user is returned
to the top-level settings view.

The top-level settings view is updated by virtue of triggering a new
settingChanged event - please use that for other settings changes.

(The css changes are expedient - some thought about organized settings
styles is needed.)