Equip PublicShareRoomModels with a sync that respects password requirement and temporarily pushes a basic auth for it.
Hook in respect for password requirements and a form to fetch it.

This all now seems to work for getting the metadata of the share room, but
it doesn't provide for the fact that the folder and file contents are also
password protected of the contents of the share room.

I will aim, next, for a model prototype on which ShareRooms, Files, and
Folders can all be based.