Use the new functional interface for getting and setting entity passwords.
Also, rectify the view finagling that's done after a valid password is
obtained, so that the share rooms root view still remains between the visit
into the passworded-shareroom's contents and the menu sheet.

That last part leaves a problem, where the share room root view is not
properly updated on resolution of the passworded-shareroom's metadata.  I'm
going to work on that next.