Implement the share room actions pop-up menu, plus the details view for public share rooms.
I've done this in a modular way, so it can be used for both "my" and public
share rooms.  The implementation for public share rooms is nearly complete.

Move FilesView.js before ShareRoomsViews.js, so stuff from the former can
be used in the latter.

Add a template for public share room details.

Also, include the name of the public share room icon as a model attribute,
for reference in the template.

(Incidentally, rearrange the tests templates list to mirror the list in the
main index.)