Add 'beenSituated' setting to public share rooms, so that we can discern those that have been fetched successfully.  We consider ones that have been established from retained records to have been fetched successfully.  This way, remembered share rooms will not be dropped if they are restored during fetch-preventing conditions, like network disconnection.
Adjust the descend_tapHandler so 'open' of public share rooms with
currently unsatisfied fetch first does a fetch, and if that still fails,
just do a notify and leave the item as an unsatisfied one.

For public share list entries that lack a name but don't have a password
pending, display a grayed ' ' "aproximately equals" sign and then the
share-id and room-key, and in the details screens, show the minimal set of
details and "Successful fetch pending" status.