Make the entire Remember Me settings entry active, and have it trigger a click on the checkbox, to have the existing change handling code do its job.
I have a question regarding this change.

I'm adding an event.preventDefault() to both the new
rememberMeSetting_tapHandler() and the already existing
rememberMe_changeHandler().  I believe, however, that the latter will never
be necessary, because the active encompassing anchor, together with the
.preventDefault() in the .rememberMeSetting_tapHandler(), will only get the
click via the .trigger('click') from the method associated with the anchor,
never directly from the UI.  Should I leave out the second
.preventDefault()?