Equip "My" share rooms with accommodations for password protection.
Similarly to the provisions for Public share rooms, we fetch the metadata
for the individual share rooms, rather than sticking with the collected
metadata reported for the account.

Also, reduce unnecessary renderings of the various elements by changing the
event which triggers render from "all" to "change".  The excessive renders
was discovered while investigating lack of update to "my" share room
entries when their passwords are provided.

There are two outstanding problems remaining with this:

1. In the root view, unlike the Public share room entries, the
   password-protected My share room entries are not updating to track
   satisfaction of authentication.

2. The share room root view My share room entries are initially presented
   with preliminary metadata, sans password-protection indications, and are
   mistakenly tappable.