Replace _.bindAll(this) with a custom version that account for the recent Underscore version's backwards-incompatible change of .bindAll() to require explicit iteration of the functions to be bound.
Use something like the method described in
https://github.com/documentcloud/underscore/commit/bf657be243a075b5e72acc8a83e6f12a564d8f55#commitcomment-2578491,
with a bit of additional effort to filter out backbone infrastructure from
the methods being this-bound.

This passes the unit tests, manual exercising of various things in my
desktop browser, and also installation on one of my devices.  (In addition
to solving #299, this also allowed me to get past the hanging tests and see
that debug:android now works, I guess resolved by the 'npm install'.)