Extend the configuration settings so that they are obtained from distinct files, and then provide the means for including brand-specific override settings - currently, the server - with a brand-specific file.
The scheme is arranged so that the brand-specific file need only specify
those settings to be overridden, leaving the rest alone. Currently, this is
not very meaningful, but this will future-proof customization code against
having to be modified as core settings accumulate.

To use it, copy www/custom_config.js into custom/brand and include entries
for the brand-specific settings.

As with all these customizations, they will only show up on the platforms
targeted by the elements.json entries.