Better incorporate backbone v 1.0 migration to doing a 'set' rather than 'reset' on fetch.
The set is already incorporating the missing items, so we don't actually
need to propagate a 'reset' in fetch.

Some views do, however, depend on getting a 'complete' event in order to
clear the "loading" indicator.  To accomplish that, we use the new
CollectionBase collection to trigger a 'complete' event, and then catch
the collection's complete event in the views to trigger a complete event
in the view.

That's all the new CollectionBase needs to do, however - no special
fetch - and the view's attention to the "reset" event is no longer
necessary.