Add our own ViewBase so we can wire in this.options, which is no longer implicitly conveyed in 1.1 - per the release notes:
  Backbone Views no longer automatically attach options passed to the
  constructor as this.options, but you can do it yourself if you
  prefer.

Having our own versions of the base elements is looking like the right
thing to do.  It will make it trivial to handle future such
accommodations (if backbone continues to churn so badly), and would
provide a basis for distilling common functionality, to reduce code
duplication.  I'm going to add an empty one for models, next.