Add recognition that the menusheet is open, and intervention so that it just closes, to all tap handlers.
In some cases, there was no 'event.preventDefault()', so I included it
only inside of the intervention, figuring it was otherwise working
without it. It might be that it should have been there, in the first
place, and should be added outside of the condition, but I opted for
minimal change from otherwise working status quo. I'd be happy to revise
this choice, make the change, if so advised.