iOS is sorted, but Android tablets still need some love
Sadly this commit also includes a hopefully temp workaround to the issue of scrolling being broken on android when using zepto 1.0.

I have put Zepto back to RC1 again until this is fixed (and added a `grunt yolo:platform` command to skip the tests until either they are fixed or we are back on Zepto 1.0).