Changed keyboard mode
At heart this is due to the keyboard coming up and fooling the view into thinking it's in landscape setting off the landscape media query in the css.

Seems that Cordova puts the android:windowSoftInputMode in the <manifest> tag instead of the <activity> tag so even though we had set it to "adjustPan" the default of "adjustResize" was being used instead.

So.. I moved it to <activity> and it fixes the problem.

BUT... it means we need to keep our eyes open for places where this messes stuff up instead of fixing it.