Complete @devgeeks switch to just using the AccountModel's notion of the current passcode and passcodeTimeout, rather than having the AccountModel manipulate the app-wide settings.
As I expected, there were a lot more changes to make to switch over to
using just the account model for passcode values, rather than having the
account model manipulate the app-wide passcode as I originally did.
However, it does yield a simpler overall scheme, so it ultimately is the
right thing to do.