New login flow...
Adds a preliminary screen that gives the option to either find out what it takes to make a new account, or to go ahead and log in.

If you choose to log in, the preliminary screen doesn't show up again until you choose it via the "X" icon on the login screen.

It also adds a menu button to the login screen for access to the Settings and ShareRooms.

Still needs copy for the "Need an account..." screen