issue 479: Start with no server set and prompt upon startup
Uses the existing server settings, it just prompts the user to change them and takes them there (and back)

Fixes #479

As a side effect, I discovered a bug in the logout after having changed servers that was leaving the app in a bad state if you tried to exit it after getting back to the login screen. I fixed that as well.

(Also sneaking in the darker blue colour change)