This turned out to be a pretty easy fix once I found the icon
I didn't like that my previous response made it sound like this was due to any inferiority of our hybrid/html5 approach. Far from it, this fix took less than 30 minutes from first google, complete with animation.

The fact is that when we first released the app, the menu indicator was not widely adopted and was only in a few of the Google apps. I wanted to implement it, but could not find an official icon for the indicator and did not want to try and create one myself.

Fixes #546