Elaborate filename colon escaping scheme only for genymotion emulator bug.
Checking this in because it's a careful solution for a complicated
problem - unfortunately, a problem that is complicated due to a bug in
the emulator, and much more than is required for the actual, android
filesystem problem.

I'm going to check in a much simpler fix for the simpler general
problem, next.

Some details of the Genymotion problem, for posterity:

Android in the emulator is responding with a FileError.INVALID_MODIFICATION_ERR
when we try to save files with names that include:

- Character entities
- asterisks
- angle brackets

None of these are a problem in other Android devices - including the
SDK emulator, or any iOS devices.