Do filename (and dirname) transforms of *only* ":".
This simplifies things, at the cost of allowing for some rare (but easy
to contrive) collisions.  Details about that below.

This means that marking as favorites files with names containing colons
or percents, or in names of containing directories, now works. I have to
work on treating stuff so viewing is handled, as well.

About collisions - since colons are translated to their character entity
equivalents, favorites of files already of that resulting form will
collide. Eg, the favorites of files with the name "with:colon" will
collide with the favorites of files with the name "with%3Acolon" in the
same dir.

As it happens, along with Cordova's .getFile() URI encoding Android
paths, anyway [1], there's no perfect solution, and this appears to be
very much the minimal disruption/maximal coverage one. Not only because
we need no new migrations, since no existing favorite would have a
different name.

[1] https://github.com/SpiderOak/SpiderOakMobileClient/issues/537#issuecomment-54383708