Fix path used for favorites when they're established in recents.
Simplify the path massage.  Specifically:

    .replace(new RegExp(model.get("url")), "");

fails to remove the file name when the removal is being done from the
recents section.

Also, use the FavoritesCollection to do that path generation, to clarify
the responsibilities.

Also, make the removal of favorite status only upon successful removal
of the file.

(There's more delegation of responsibilities to other objects that could
clarify and simplify the process a lot. I'm going to do a bit of that,
but it's now working.)