Refactor favorites removal according to responsibilities.
This continues to work solidly, regardless of where the the removal is
initiated.

We need to instruct the testers about interventions to take to clear
disruptive residue from the prior problems. I'll add that info to the PR.

The file deletion should eventually also go into the favorites, not the
view, but that will entail deeper changes to the collections and models
than I want to broach for this issue.