Hook in localization machinery.
I needed to be able to see if things were working right, so I did
something along the lines of
https://github.com/mclear/NFC_Ring_Control/blob/df8db31dd1683b04422c106a1484637629b4c88f/www/js/nfcRing/ui.js#L125-L135

The operating sequence seems to work, but I'm far from sure this is
optimal. I am sure something needs to be done to stub translations for
the tests, to at least quell the "no translations" warnings, but I need
to return to bracketing and cataloging the strings before looking at
that.

(Interoperation with our brand-string provisions turns out to be no
problem at all, due to html10n's template parameter provision.)