Imported locales, moment.js and forked html10n.js
i18n.json: This commit changes the i18n.json file to use imports for the various locales.

moment.js: Updated moment to v2.8.0 (better locale support) and changed the Gruntfile to use the moment file with the locales baked in.

html10n.js: There were problems in html10n.js with using imports in i18n.json that had been fixed in the NFC Ring app we are using as a template. I forked html10n.js and modified it with those hacks. I also changed bower to use the fork (it's also using a branch on the fork in case we want to feed fixes back to html10n.js)..