Apply some overlooked translations, while looking at action menus.
This adds to our list of phrases not sent to the translators, for the
initial Latin American Spanish pass:

- Share file
- Open in

Having blithely missed these suggests we need another review, with
translated strings highlighted so that missed source phrases can be
noticed. The highlighting could easily be done by doing a build while
the translation routine is rigged to bracket the translations with
distinctive characters.