Workaround html10n.build() bizarre behavior with arrays that include "es-es".
Stepping through reveals that html10n.build() resolves langs list
argument ["es-es", "es", nil, "en"] to ["es-es", "es", "es", "es"], or
something to that effect. This explains the lack of fallthrough for
es. The workaround avoids that sensitivity.

The real fix is to see stop html10n.build() from doing that. That would
be easier if we knew what it was supposed to be doing. Sigh.