Use the English version of "Zero-Knowledge" in the Spanish translation.
I haven't got preservation of the English string working using
bracketing, so for now we have to manually switch back the translated
phrases. I seem to recall that the html10n.get() code would use
translations for the substitutions, but also recall noticing that wasn't
working. Have to fix it, if it's there, or else just use trivial
identity replacements.