Add missing bracketing variant for "Not '[[Zero-Knowledge]]'"
In some cases the square bracketed version was missing, in some cases
the curly one was. Almost all were missing one or the other - which
breaks either a template that uses the fragment or some code.

This only increases the urgency of normalizing to all source string
bracketing using squares. This will reduce confusion, both for our own
implementation and for our translators, for whom bracketing is
oblique enough without this wrinkle.