Start username phase-out, by asking for/presenting "Email/Username"
In the settings account view, we were already presenting the "Login
name" if it differed from the username. Now the logic is simpler - we
remove the condition and just present what was in this Login name field,
which is whatever the user logged in with.  To my current knowledge,
that's the best we can do.

The layout constraints seem ok, reasonably truncating the fields so the
actual data gets space.