Fix server-change facility by Providing missing l10n parameters.
I used this regex search across the sources to look for other instances
where apparently no parameters were provided:

  find . -name "*.js" -print0 | xargs -0 grep -n 'qq("[^"]*\[' | grep -v ':,'

(The lack of parameters is inferred by the lack of a trailing "," no the
key string. Not fully conclusive, this is a good check.)

We should change html10n code to emit a console error message when the
parameters are missing, rather than crashing the app.