Continuing translations adjustments, for simplest baseline.
1. Use only "[["/"]]" for all source strings brackets.
2. Bracket all literals not intended to be translated.

Both measures simplify the instructions, reducing opportunities for
confusion and mistakes.

Item #2 also the opportunity to tailor some translations to present both
the English, literal version while also providing room for a
translation-specific qualifier, in cases - particularly for languages
that use very different character sets - where an additional cue
accompanying the string, in the local language, is called for.