Make en-GB the base fall-through, and bracket a non-translated string.
Many more localities use en-GB than en-US, among the localities we are
likely to cover, so we're making en-GB the base fall-through. This
reduces the amount of configuration necessary, while en-US is still easy
to cover. (en-US actually accounts for a major portion of the number of
English speakers, but not the number of localities, at least those with
we're initially concerned.)

Also, start to bracket non-translated strings to distinguish them for
non-translation, and to facilitate ornamenting (eg, with added local-language
cue) in those languages where ornamenting is warranted. More such
bracketing to go.