Localize Blue aboutSpiderOakViewTemplate.html template.
The only new string is an exact portion of an existing translated
string, so I've made the substitution in all the languages. I'm pretty
confident that the grammar works fine.

Also, the only content differences between the Blue and Orange versions
of the needAnAccountViewTemplate.html are unimportant, so we're removing
the Blue variant.