Implement a version of readyFinisher which prevents misfires.
This version ensures that the exact, specified contingencies are
accounted for. The previous version would misfire if, eg, one of the
callers unexpectedly runs repeatedly.

This version is generalized, so it can be applied to similar situations
where coordination is needed.

These two benefits account for the immediate advantages that the
promise-based implementation had, so I now prefer this one.

Provided we have tests for readyFinisher, up next.