Upgrade chai and sinon-chai to accommodate chai-as-promised requirements.
chai-as-promised npm dependencies require a more recent version of chai
than we were using, but that slipped by, somehow.

To exercise the dependencies I did an npm install with an empty
node_modules directory, and then with an empty bower_components
directory, to verify that both succeed and that `grunt test` (and a
`grunt debug:ios`) work fine.